Got types of srcFileNum and newFileNum reversed.

+++ unix/tclXunixDup.c
@@ -76,8 +76,8 @@ TclXOSDupChannel(Tcl_Interp *interp, Tcl_Channel srcChannel, int mode, char *tar
     ClientData handle;
     const Tcl_ChannelType *channelType;
     Tcl_Channel newChannel = NULL;
-    int srcFileNum;
-    intptr_t newFileNum = -1;
+    intptr_t srcFileNum;
+    int newFileNum = -1;
 
     /*
      * On Unix, the channels we can dup share the same file for the read and
